requirements.txt
groq>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
flask>=3.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tqdm>=4.66.0
